Fix usages of modalService.open that don't catch the result promise
