--- app-frontend/src/app/components/exports/exportItem/exportItem.component.js
@@ -1,13 +1,0 @@
-// Component code
-import exportTpl from './exportItem.html';
-
-const rfExportItem = {
-    templateUrl: exportTpl,
-    transclude: true,
-    controller: 'ExportItemController',
-    bindings: {
-        export: '<'
-    }
-};
-
-export default rfExportItem;
