--- app-frontend/src/app/components/projects/projectCreateModal/projectCreateModal.component.js
@@ -1,14 +1,0 @@
-import projectCreateModalTpl from './projectCreateModal.html';
-
-const rfProjectCreateModal = {
-    templateUrl: projectCreateModalTpl,
-    bindings: {
-        close: '&',
-        dismiss: '&',
-        modalInstance: '<',
-        resolve: '<'
-    },
-    controller: 'ProjectCreateModalController'
-};
-
-export default rfProjectCreateModal;
