+++ app-frontend/src/app/components/scenes/sceneDetailModal/sceneDetailModal.module.js
@@ -88,7 +88,7 @@ class SceneDetailModalController {
             resolve: {
                 scene: () => this.scene
             }
-        });
+        }).result.catch(() => {});
     }
 
     getSceneBounds() {
