--- app-frontend/src/app/pages/imports/datasources/datasources.controller.js
@@ -1,6 +1,0 @@
-class DatasourcesController {
-    constructor() { }
-
-}
-
-export default DatasourcesController;
