--- app-frontend/src/app/pages/imports/imports.controller.js
@@ -1,21 +1,0 @@
-/* global BUILDCONFIG */
-
-class ImportsController {
-    constructor(authService, modalService) {
-        'ngInject';
-        this.authService = authService;
-        this.modalService = modalService;
-    }
-
-    $onInit() {
-        this.BUILDCONFIG = BUILDCONFIG;
-    }
-
-    openCreateDatasourceModal() {
-        this.modalService.open({
-            component: 'rfDatasourceCreateModal'
-        });
-    }
-}
-
-export default ImportsController;
