+++ app-frontend/src/app/pages/imports/imports.module.js
@@ -1,5 +1,24 @@
-import ImportsController from './imports.controller.js';
+/* global BUILDCONFIG */
 import autoInject from '_appRoot/autoInject';
+
+
+class ImportsController {
+    constructor(authService, modalService) {
+        'ngInject';
+        this.authService = authService;
+        this.modalService = modalService;
+    }
+
+    $onInit() {
+        this.BUILDCONFIG = BUILDCONFIG;
+    }
+
+    openCreateDatasourceModal() {
+        this.modalService.open({
+            component: 'rfDatasourceCreateModal'
+        }).result.catch(() => {});
+    }
+}
 
 const ImportsModule = angular.module('pages.imports', []);
