+++ app-frontend/src/app/pages/lab/browse/analyses/analyses.js
@@ -171,7 +171,7 @@ class LabBrowseAnalysesController {
             resolve: {
                 analysis: () => analysis
             }
-        });
+        }).result.catch(() => {});
     }
 
     checkAnalysesExports(analysesList) {
