+++ app-frontend/src/app/pages/lab/browse/templates/templates.js
@@ -58,7 +58,7 @@ class LabBrowseTemplatesController {
     openTemplateCreateModal() {
         this.modalService.open({
             component: 'rfTemplateCreateModal'
-        });
+        }).result.catch(() => {});
     }
 
     onTemplateDelete(templateId) {
